+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/components/style/StyleFactory.kt
@@ -2,7 +2,6 @@ package com.revenuecat.purchases.ui.revenuecatui.components.style
 
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.unit.dp
-import com.revenuecat.purchases.paywalls.components.BadgeComponent
 import com.revenuecat.purchases.paywalls.components.ButtonComponent
 import com.revenuecat.purchases.paywalls.components.ImageComponent
 import com.revenuecat.purchases.paywalls.components.PackageComponent
@@ -29,6 +28,7 @@ import com.revenuecat.purchases.ui.revenuecatui.components.toPresentedOverrides
 import com.revenuecat.purchases.ui.revenuecatui.errors.PaywallValidationError
 import com.revenuecat.purchases.ui.revenuecatui.helpers.NonEmptyList
 import com.revenuecat.purchases.ui.revenuecatui.helpers.Result
+import com.revenuecat.purchases.ui.revenuecatui.helpers.getOrThrow
 import com.revenuecat.purchases.ui.revenuecatui.helpers.map
 import com.revenuecat.purchases.ui.revenuecatui.helpers.mapError
 import com.revenuecat.purchases.ui.revenuecatui.helpers.mapOrAccumulate
@@ -53,7 +53,6 @@ internal class StyleFactory(
         actionHandler: suspend (PaywallAction) -> Unit,
     ): Result<ComponentStyle, NonEmptyList<PaywallValidationError>> =
         when (component) {
-            is BadgeComponent -> TODO("BadgeComponentStyle is not yet implemented")
             is ButtonComponent -> createButtonComponentStyle(component, actionHandler)
             is ImageComponent -> TODO("ImageComponentStyle is not yet implemented.")
             is PackageComponent -> TODO("PackageComponentStyle is not yet implemented.")
@@ -128,6 +127,16 @@ internal class StyleFactory(
             state = componentState,
         )?.partial
 
+        val badge = partial?.badge ?: component.badge
+        val badgeStyle = badge?.let {
+            BadgeStyle(
+                // TODO Need to map these errors instead of throwing.
+                stackStyle = createStackComponentStyle(it.stack).getOrThrow(),
+                style = it.style,
+                alignment = it.alignment,
+            )
+        }
+
         StackComponentStyle(
             visible = partial?.visible ?: true,
             children = children,
@@ -140,6 +149,7 @@ internal class StyleFactory(
             shape = (partial?.shape ?: component.shape)?.toShape() ?: DEFAULT_SHAPE,
             border = partial?.border ?: component.border,
             shadow = partial?.shadow ?: component.shadow,
+            badge = badgeStyle,
         )
     }
